moved db filling functions to a different file.
